% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEWorkshop-package.R
\docType{package}
\name{iSEEWorkshop-package}
\alias{iSEEWorkshop}
\alias{iSEEWorkshop-package}
\title{iSEEWorkshop: A workshop on the iSEE package}
\description{
This workshop covers basic usage of iSEE for 
    interactive visualization of SummarizedExperiment objects.
}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk} (0000-0003-3899-3872)

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (0000-0003-3252-7758)
  \item Charlotte Soneson \email{charlottesoneson@gmail.com} (0000-0003-3833-2169)
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (0000-0002-3564-4813)
}

}
\keyword{internal}
